Split autonomous logic and routines into seperate files
